pyproject.toml
[project]
name = "myvibecodingagent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "openai==2.29.0",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
